+++ test/src/androidTest/kotlin/io/realm/shared/ImportTests.kt
@@ -44,7 +44,7 @@ class ImportTests {
         tmpDir = PlatformUtils.createTempDir()
         val configuration =
             RealmConfiguration(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Parent::class, Child::class, Sample::class))
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
     }
 
     @AfterTest