+++ test/src/androidTest/kotlin/io/realm/shared/NullabilityTests.kt
@@ -35,7 +35,7 @@ class NullabilityTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Nullability::class))
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
     }
 
     @AfterTest