+++ test/src/androidTest/kotlin/io/realm/shared/RealmObjectTests.kt
@@ -50,7 +50,7 @@ class RealmObjectTests : RealmLifeCycleTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Parent::class, Child::class))
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         parent = realm.writeBlocking { copyToRealm(Parent()) }
     }