+++ test/src/androidTest/kotlin/io/realm/shared/RealmTests.kt
@@ -130,10 +130,10 @@ class RealmTests {
             path = "$tmpDir/exceed-versions.realm",
             schema = setOf(Parent::class, Child::class)
         ).maxNumberOfActiveVersions(1).build()
-        realm = Realm.open(config)
+        realm = Realm(config)
         // Pin the version, so when starting a new transaction on the first Realm,
         // we don't release older versions.
-        val otherRealm = Realm.open(config)
+        val otherRealm = Realm(config)
 
         try {
             // FIXME Should be IllegalStateException
@@ -223,7 +223,7 @@ class RealmTests {
         // Ensure that all writes are actually committed
         realm.close()
         assertTrue(realm.isClosed())
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         assertEquals(10, realm.objects(Parent::class).size)
     }
 
@@ -263,7 +263,7 @@ class RealmTests {
         realm.close()
         assertTrue(realm.isClosed())
 
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         assertEquals(1, realm.objects(Parent::class).size)
     }
 
@@ -285,7 +285,7 @@ class RealmTests {
         writeStarted.lock()
         realm.close()
         assert(write.await() is RuntimeException)
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         assertEquals(0, realm.objects<Parent>().size)
     }
 
@@ -322,7 +322,7 @@ class RealmTests {
         // Ensure that write is not committed
         realm.close()
         assertTrue(realm.isClosed())
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         // This assertion doesn't hold on MacOS as all code executes on the same thread as the
         // dispatcher is a run loop on the local thread, thus, the main flow is not picked up when
         // the mutex is unlocked. Doing so would require the write block to be able to suspend in
@@ -356,7 +356,7 @@ class RealmTests {
         // Ensure that only one write is actually committed
         realm.close()
         assertTrue(realm.isClosed())
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
         // This assertion doesn't hold on MacOS as all code executes on the same thread as the
         // dispatcher is a run loop on the local thread, thus, the main flow is not picked up when
         // the mutex is unlocked. Doing so would require the write block to be able to suspend in