+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmNotificationsTests.kt
@@ -27,7 +27,7 @@ class RealmNotificationsTests : NotificationTests {
         tmpDir = PlatformUtils.createTempDir()
         configuration =
             RealmConfiguration(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
     }
 
     @AfterTest