+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmResultsNotificationsTests.kt
@@ -31,7 +31,7 @@ class RealmResultsNotificationsTests : NotificationTests {
         tmpDir = PlatformUtils.createTempDir()
         configuration =
             RealmConfiguration(path = "$tmpDir/default.realm", schema = setOf(Sample::class))
-        realm = Realm.open(configuration)
+        realm = Realm(configuration)
     }
 
     @AfterTest